A second <router-outlet> in a features module and Angular Child routes
Caveats:

- everything is first loaded from the server to the client (no lazy-loading, but eager-loading!)

- with PathLocationStrategy, when you enter the direct URL of the route in the browser’s address bar, it still makes a request to the server, which won’t find the resource (and rightly so) named as your route. This will cause a 404 error.
Can be solved by configuring the web server to do a redirect to index.html of the app in cases when a requested resource isn’t found. This will put the Angular app back in control, and the route will be properly resolved. The Angular CLI development server is already configured for redirects. But then the wildcard route (path: '**') can't be used to inform users that there is a 404 failure...